// projects/iblopay-admin/src/app/modules/services-publics/models/service-public.model.d.ts
export interface Institution {
    id: number;
    nom: string;
    code: string;
    description: string;
    type: 'ARCT' | 'OBM' | 'OBPE' | 'OBUHA' | 'ARCA' | 'MINEDUC' | 'MININTER' | 'MINJUST' | 'MINREX' | 'ABREMA' | 'OHP' | 'LONA' | 'OTRACO' | 'AUTRE';
    actif: boolean;
    dateCreation: Date;
    dateModification?: Date;
    serviceId: number;
    contact?: string;
    email?: string;
    telephone?: string;
    adresse?: string;
    statistiques?: InstitutionStatistiques;
}
export interface InstitutionStatistiques {
    totalTypesRNF: number;
    totalPaiements: number;
    montantTotalPaiements: number;
}